refactor(ffi): extract update manager and release entrypoints
