+++ crates/surge-ffi/src/lib.rs
@@ -16,7 +16,9 @@
 
 mod context;
 mod handles;
+mod releases;
 mod shared;
+mod update;
 mod utils;
 
 use std::collections::BTreeMap;
@@ -28,19 +30,25 @@ use surge_core::diff::wrapper::{bsdiff_buffers, bspatch_buffers};
 use surge_core::lock::mutex::DistributedMutex;
 use surge_core::pack::builder::PackBuilder;
 use surge_core::supervisor::state::{supervisor_pid_file, supervisor_stop_file, write_restart_args};
-use surge_core::update::manager::{ProgressInfo, UpdateManager};
 
 pub use crate::context::{
     surge_config_set_lock_server, surge_config_set_resource_budget, surge_config_set_storage, surge_context_create,
     surge_context_destroy, surge_context_last_error,
 };
-use crate::handles::{
-    ReleaseEntryFfi, SurgeContextHandle, SurgePackContextHandle, SurgeReleasesInfoHandle, SurgeUpdateManagerHandle,
+use crate::handles::{SurgeContextHandle, SurgePackContextHandle};
+pub use crate::releases::{
+    surge_release_channel, surge_release_full_size, surge_release_is_genesis, surge_release_version,
+    surge_releases_count, surge_releases_destroy,
 };
 use crate::shared::{
-    ProgressBridge, SURGE_CANCELLED, SURGE_ERROR, SURGE_NOT_FOUND, SURGE_OK, SURGE_PHASE_CHECK, SURGE_PHASE_DOWNLOAD,
-    SurgeEventCallback, SurgeProgressCallback, catch_ffi, clear_shared_error, collect_argv, cstr_to_string,
-    libc_malloc, make_pack_progress, set_ctx_error, set_shared_error, try_index, try_len, try_len_allow_zero,
+    SURGE_CANCELLED, SURGE_ERROR, SURGE_OK, SURGE_PHASE_CHECK, SURGE_PHASE_DOWNLOAD, SurgeEventCallback,
+    SurgeProgressCallback, catch_ffi, clear_shared_error, collect_argv, cstr_to_string, libc_malloc,
+    make_pack_progress, set_ctx_error, set_shared_error, try_len, try_len_allow_zero,
+};
+pub use crate::update::{
+    surge_update_check, surge_update_download_and_apply, surge_update_manager_create, surge_update_manager_destroy,
+    surge_update_manager_set_channel, surge_update_manager_set_current_version,
+    surge_update_manager_set_release_retention_limit,
 };
 use crate::utils::{lock_recover, to_lossy_cstring};
 
@@ -89,412 +97,6 @@ pub struct SurgeBspatchCtxFfi {
     pub patch: *const u8,
     pub patch_size: i64,
     pub status: i32,
-}
-
-// =========================================================================
-//  3. Update manager (4 functions)
-// =========================================================================
-
-/// Create an update manager bound to a specific application.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_update_manager_create(
-    ctx: *mut SurgeContextHandle,
-    app_id: *const c_char,
-    current_version: *const c_char,
-    channel: *const c_char,
-    install_dir: *const c_char,
-) -> *mut SurgeUpdateManagerHandle {
-    if ctx.is_null() {
-        return ptr::null_mut();
-    }
-
-    let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        // SAFETY: `ctx` is checked non-null above and refers to a valid
-        // handle for the duration of this call.
-        let handle = unsafe { &*ctx };
-        handle.clear_last_error();
-
-        // SAFETY: string pointer inputs follow the FFI contract described in
-        // this API; nulls map to empty strings.
-        let (app_id_s, version_s, channel_s, install_s) = unsafe {
-            (
-                cstr_to_string(app_id),
-                cstr_to_string(current_version),
-                cstr_to_string(channel),
-                cstr_to_string(install_dir),
-            )
-        };
-
-        if app_id_s.is_empty() || version_s.is_empty() || channel_s.is_empty() || install_s.is_empty() {
-            let e =
-                surge_core::error::SurgeError::Config("app_id, version, channel, and install_dir are required".into());
-            set_ctx_error(handle, &e);
-            return ptr::null_mut();
-        }
-
-        let mgr = Box::new(SurgeUpdateManagerHandle {
-            ctx: handle.ctx.clone(),
-            runtime: handle.runtime.clone(),
-            last_error: handle.shared_last_error.clone(),
-            app_id: app_id_s,
-            current_version: version_s,
-            channel: channel_s,
-            release_retention_limit: 1,
-            install_dir: install_s,
-        });
-
-        Box::into_raw(mgr)
-    }));
-
-    result.unwrap_or(ptr::null_mut())
-}
-
-/// Destroy an update manager.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_update_manager_destroy(mgr: *mut SurgeUpdateManagerHandle) {
-    if !mgr.is_null() {
-        let _ = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            drop(unsafe { Box::from_raw(mgr) });
-        }));
-    }
-}
-
-/// Change the update channel used by this manager.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_update_manager_set_channel(
-    mgr: *mut SurgeUpdateManagerHandle,
-    channel: *const c_char,
-) -> i32 {
-    if mgr.is_null() {
-        return SURGE_ERROR;
-    }
-
-    catch_ffi(std::panic::AssertUnwindSafe(|| {
-        // SAFETY: `mgr` is checked non-null above.
-        let mgr_ref = unsafe { &mut *mgr };
-        clear_shared_error(&mgr_ref.ctx, &mgr_ref.last_error);
-
-        // SAFETY: `channel` follows the nullable C string contract.
-        let channel_s = unsafe { cstr_to_string(channel) };
-        let channel_s = channel_s.trim().to_string();
-        if channel_s.is_empty() {
-            let e = surge_core::error::SurgeError::Config("channel is required".into());
-            return set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e);
-        }
-
-        mgr_ref.channel = channel_s;
-        SURGE_OK
-    }))
-}
-
-/// Change the current installed version baseline used for update checks.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_update_manager_set_current_version(
-    mgr: *mut SurgeUpdateManagerHandle,
-    current_version: *const c_char,
-) -> i32 {
-    if mgr.is_null() {
-        return SURGE_ERROR;
-    }
-
-    catch_ffi(std::panic::AssertUnwindSafe(|| {
-        // SAFETY: `mgr` is checked non-null above.
-        let mgr_ref = unsafe { &mut *mgr };
-        clear_shared_error(&mgr_ref.ctx, &mgr_ref.last_error);
-
-        // SAFETY: `current_version` follows the nullable C string contract.
-        let version_s = unsafe { cstr_to_string(current_version) };
-        let version_s = version_s.trim().to_string();
-        if version_s.is_empty() {
-            let e = surge_core::error::SurgeError::Config("current_version is required".into());
-            return set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e);
-        }
-
-        mgr_ref.current_version = version_s;
-        SURGE_OK
-    }))
-}
-
-/// Change the number of old app versions retained after successful updates.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_update_manager_set_release_retention_limit(
-    mgr: *mut SurgeUpdateManagerHandle,
-    release_retention_limit: c_int,
-) -> i32 {
-    if mgr.is_null() {
-        return SURGE_ERROR;
-    }
-
-    catch_ffi(std::panic::AssertUnwindSafe(|| {
-        // SAFETY: `mgr` is checked non-null above.
-        let mgr_ref = unsafe { &mut *mgr };
-        clear_shared_error(&mgr_ref.ctx, &mgr_ref.last_error);
-
-        if release_retention_limit < 0 {
-            let e = surge_core::error::SurgeError::Config("release_retention_limit cannot be negative".into());
-            return set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e);
-        }
-
-        mgr_ref.release_retention_limit = usize::try_from(release_retention_limit).unwrap_or(usize::MAX);
-        SURGE_OK
-    }))
-}
-
-/// Check for available updates.
-///
-/// Returns `SURGE_OK` if updates are available, `SURGE_NOT_FOUND` if up-to-date.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_update_check(
-    mgr: *mut SurgeUpdateManagerHandle,
-    info: *mut *mut SurgeReleasesInfoHandle,
-) -> i32 {
-    if mgr.is_null() || info.is_null() {
-        return SURGE_ERROR;
-    }
-
-    catch_ffi(std::panic::AssertUnwindSafe(|| {
-        // SAFETY: `mgr`/`info` are checked non-null above. The out pointer is
-        // cleared first to avoid leaking stale values on early exits.
-        let mgr_ref = unsafe {
-            *info = ptr::null_mut();
-            &*mgr
-        };
-        clear_shared_error(&mgr_ref.ctx, &mgr_ref.last_error);
-
-        if mgr_ref.ctx.is_cancelled() {
-            return SURGE_CANCELLED;
-        }
-
-        let mut update_mgr = match UpdateManager::new(
-            mgr_ref.ctx.clone(),
-            &mgr_ref.app_id,
-            &mgr_ref.current_version,
-            &mgr_ref.channel,
-            &mgr_ref.install_dir,
-        ) {
-            Ok(m) => m,
-            Err(e) => return set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e),
-        };
-        update_mgr.set_release_retention_limit(mgr_ref.release_retention_limit);
-
-        let result = mgr_ref.runtime.block_on(update_mgr.check_for_updates());
-
-        match result {
-            Ok(Some(update_info)) => {
-                let ffi_releases: Vec<ReleaseEntryFfi> = update_info
-                    .available_releases
-                    .iter()
-                    .map(|r| ReleaseEntryFfi {
-                        version: r.version.clone(),
-                        channel: mgr_ref.channel.clone(),
-                        full_size: r.full_size,
-                        is_genesis: r.is_genesis,
-                    })
-                    .collect();
-
-                let mut releases_handle = Box::new(SurgeReleasesInfoHandle {
-                    releases: ffi_releases,
-                    cached_strings: Vec::new(),
-                    update_info: Some(update_info),
-                });
-                releases_handle.cache_strings();
-
-                // SAFETY: `info` is a valid out pointer checked above.
-                unsafe { *info = Box::into_raw(releases_handle) };
-                SURGE_OK
-            }
-            Ok(None) => {
-                let releases_handle = Box::new(SurgeReleasesInfoHandle {
-                    releases: Vec::new(),
-                    cached_strings: Vec::new(),
-                    update_info: None,
-                });
-                // SAFETY: `info` is a valid out pointer checked above.
-                unsafe { *info = Box::into_raw(releases_handle) };
-                SURGE_NOT_FOUND
-            }
-            Err(e) => set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e),
-        }
-    }))
-}
-
-/// Download and apply an update described by `info`.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_update_download_and_apply(
-    mgr: *mut SurgeUpdateManagerHandle,
-    info: *const SurgeReleasesInfoHandle,
-    progress_cb: SurgeProgressCallback,
-    user_data: *mut c_void,
-) -> i32 {
-    if mgr.is_null() || info.is_null() {
-        return SURGE_ERROR;
-    }
-
-    catch_ffi(std::panic::AssertUnwindSafe(|| {
-        let mgr_ref = unsafe { &*mgr };
-        clear_shared_error(&mgr_ref.ctx, &mgr_ref.last_error);
-
-        if mgr_ref.ctx.is_cancelled() {
-            return SURGE_CANCELLED;
-        }
-
-        let info_ref = unsafe { &*info };
-
-        let update_info = match info_ref.update_info.as_ref() {
-            Some(ui) => ui,
-            None => {
-                let e = surge_core::error::SurgeError::Update("No update info available".into());
-                return set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e);
-            }
-        };
-
-        if update_info.available_releases.is_empty() {
-            let e = surge_core::error::SurgeError::Update("No releases to apply".into());
-            return set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e);
-        }
-
-        let mut update_mgr = match UpdateManager::new(
-            mgr_ref.ctx.clone(),
-            &mgr_ref.app_id,
-            &mgr_ref.current_version,
-            &mgr_ref.channel,
-            &mgr_ref.install_dir,
-        ) {
-            Ok(m) => m,
-            Err(e) => return set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e),
-        };
-        update_mgr.set_release_retention_limit(mgr_ref.release_retention_limit);
-
-        // Map core ProgressInfo → SurgeProgressFfi via the C callback.
-        // Handle Some/None explicitly to give the compiler a concrete type
-        // for the generic F in download_and_apply<F: Send + Sync>.
-        let result = if let Some(cb) = progress_cb {
-            let bridge = ProgressBridge {
-                cb,
-                user_data: user_data as usize,
-            };
-            let progress_fn = move |pi: ProgressInfo| {
-                bridge.invoke(&pi);
-            };
-            mgr_ref
-                .runtime
-                .block_on(update_mgr.download_and_apply(update_info, Some(progress_fn)))
-        } else {
-            mgr_ref
-                .runtime
-                .block_on(update_mgr.download_and_apply(update_info, None::<fn(ProgressInfo)>))
-        };
-
-        match result {
-            Ok(()) => SURGE_OK,
-            Err(e) => set_shared_error(&mgr_ref.ctx, &mgr_ref.last_error, &e),
-        }
-    }))
-}
-
-// =========================================================================
-//  4. Release-info accessors (6 functions)
-// =========================================================================
-
-/// Return the number of releases in `info`.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_releases_count(info: *const SurgeReleasesInfoHandle) -> i32 {
-    if info.is_null() {
-        return 0;
-    }
-
-    let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        let h = unsafe { &*info };
-        i32::try_from(h.releases.len()).unwrap_or(i32::MAX)
-    }));
-
-    result.unwrap_or(0)
-}
-
-/// Free a releases-info structure returned by `surge_update_check`.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_releases_destroy(info: *mut SurgeReleasesInfoHandle) {
-    if !info.is_null() {
-        let _ = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            drop(unsafe { Box::from_raw(info) });
-        }));
-    }
-}
-
-/// Return the version string for release at `index`.
-///
-/// The returned pointer is valid for the lifetime of the `info` handle.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_release_version(info: *const SurgeReleasesInfoHandle, index: i32) -> *const c_char {
-    if info.is_null() {
-        return ptr::null();
-    }
-
-    let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        let h = unsafe { &*info };
-        let Some(idx) = try_index(index, h.cached_strings.len()) else {
-            return ptr::null();
-        };
-        h.cached_strings[idx].0.as_ptr()
-    }));
-
-    result.unwrap_or(ptr::null())
-}
-
-/// Return the channel string for release at `index`.
-///
-/// The returned pointer is valid for the lifetime of the `info` handle.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_release_channel(info: *const SurgeReleasesInfoHandle, index: i32) -> *const c_char {
-    if info.is_null() {
-        return ptr::null();
-    }
-
-    let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        let h = unsafe { &*info };
-        let Some(idx) = try_index(index, h.cached_strings.len()) else {
-            return ptr::null();
-        };
-        h.cached_strings[idx].1.as_ptr()
-    }));
-
-    result.unwrap_or(ptr::null())
-}
-
-/// Return the full-package size in bytes for release at `index`.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_release_full_size(info: *const SurgeReleasesInfoHandle, index: i32) -> i64 {
-    if info.is_null() {
-        return 0;
-    }
-
-    let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        let h = unsafe { &*info };
-        let Some(idx) = try_index(index, h.releases.len()) else {
-            return 0;
-        };
-        h.releases[idx].full_size
-    }));
-
-    result.unwrap_or(0)
-}
-
-/// Return non-zero if release at `index` is a genesis (initial) release.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn surge_release_is_genesis(info: *const SurgeReleasesInfoHandle, index: i32) -> i32 {
-    if info.is_null() {
-        return 0;
-    }
-
-    let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        let h = unsafe { &*info };
-        let Some(idx) = try_index(index, h.releases.len()) else {
-            return 0;
-        };
-        i32::from(h.releases[idx].is_genesis)
-    }));
-
-    result.unwrap_or(0)
 }
 
 // =========================================================================
@@ -1167,155 +769,8 @@ pub unsafe extern "C" fn surge_cancel(ctx: *mut SurgeContextHandle) -> i32 {
 
 #[cfg(test)]
 mod tests {
-    use std::ffi::{CStr, CString};
-
     use super::*;
-
-    #[test]
-    fn manager_set_channel_updates_context_last_error() {
-        let ctx = surge_context_create();
-        assert!(!ctx.is_null());
-
-        let app_id = CString::new("demo").unwrap();
-        let version = CString::new("1.0.0").unwrap();
-        let channel = CString::new("stable").unwrap();
-        let install_dir = CString::new("/tmp/demo").unwrap();
-
-        let mgr = unsafe {
-            surge_update_manager_create(
-                ctx,
-                app_id.as_ptr(),
-                version.as_ptr(),
-                channel.as_ptr(),
-                install_dir.as_ptr(),
-            )
-        };
-        assert!(!mgr.is_null());
-
-        let empty = CString::new("").unwrap();
-        let rc = unsafe { surge_update_manager_set_channel(mgr, empty.as_ptr()) };
-        assert_ne!(rc, SURGE_OK);
-
-        let last = unsafe { surge_context_last_error(ctx) };
-        assert!(!last.is_null());
-        let msg = unsafe { CStr::from_ptr((*last).message) }
-            .to_str()
-            .unwrap_or_default()
-            .to_string();
-        assert!(msg.contains("channel"));
-
-        unsafe {
-            surge_update_manager_destroy(mgr);
-            surge_context_destroy(ctx);
-        }
-    }
-
-    #[test]
-    fn manager_set_release_retention_limit_updates_context_last_error() {
-        let ctx = surge_context_create();
-        assert!(!ctx.is_null());
-
-        let app_id = CString::new("demo").unwrap();
-        let version = CString::new("1.0.0").unwrap();
-        let channel = CString::new("stable").unwrap();
-        let install_dir = CString::new("/tmp/demo").unwrap();
-
-        let mgr = unsafe {
-            surge_update_manager_create(
-                ctx,
-                app_id.as_ptr(),
-                version.as_ptr(),
-                channel.as_ptr(),
-                install_dir.as_ptr(),
-            )
-        };
-        assert!(!mgr.is_null());
-
-        let rc = unsafe { surge_update_manager_set_release_retention_limit(mgr, -1) };
-        assert_ne!(rc, SURGE_OK);
-
-        let last = unsafe { surge_context_last_error(ctx) };
-        assert!(!last.is_null());
-        let msg = unsafe { CStr::from_ptr((*last).message) }
-            .to_str()
-            .unwrap_or_default()
-            .to_string();
-        assert!(msg.contains("release_retention_limit"));
-
-        unsafe {
-            surge_update_manager_destroy(mgr);
-            surge_context_destroy(ctx);
-        }
-    }
-
-    #[test]
-    fn manager_remains_usable_after_context_destroy() {
-        let ctx = surge_context_create();
-        assert!(!ctx.is_null());
-
-        let app_id = CString::new("demo").unwrap();
-        let version = CString::new("1.0.0").unwrap();
-        let channel = CString::new("stable").unwrap();
-        let install_dir = CString::new("/tmp/demo").unwrap();
-
-        let mgr = unsafe {
-            surge_update_manager_create(
-                ctx,
-                app_id.as_ptr(),
-                version.as_ptr(),
-                channel.as_ptr(),
-                install_dir.as_ptr(),
-            )
-        };
-        assert!(!mgr.is_null());
-
-        unsafe { surge_context_destroy(ctx) };
-
-        let test_channel = CString::new("test").unwrap();
-        let rc = unsafe { surge_update_manager_set_channel(mgr, test_channel.as_ptr()) };
-        assert_eq!(rc, SURGE_OK);
-
-        unsafe { surge_update_manager_destroy(mgr) };
-    }
-
-    #[test]
-    fn update_check_clears_output_pointer_on_failure() {
-        let ctx = surge_context_create();
-        assert!(!ctx.is_null());
-
-        let app_id = CString::new("demo").unwrap();
-        let version = CString::new("1.0.0").unwrap();
-        let channel = CString::new("stable").unwrap();
-        let install_dir = CString::new("/tmp/demo").unwrap();
-
-        let mgr = unsafe {
-            surge_update_manager_create(
-                ctx,
-                app_id.as_ptr(),
-                version.as_ptr(),
-                channel.as_ptr(),
-                install_dir.as_ptr(),
-            )
-        };
-        assert!(!mgr.is_null());
-
-        let stale = Box::new(SurgeReleasesInfoHandle {
-            releases: Vec::new(),
-            cached_strings: Vec::new(),
-            update_info: None,
-        });
-        let mut info_ptr = Box::into_raw(stale);
-        assert!(!info_ptr.is_null());
-
-        let rc = unsafe { surge_update_check(mgr, &mut info_ptr) };
-        assert_ne!(rc, SURGE_OK);
-        assert!(info_ptr.is_null());
-
-        unsafe {
-            surge_update_manager_destroy(mgr);
-            surge_context_destroy(ctx);
-        }
-    }
+    use std::ffi::{CStr, CString};
 
     #[test]
     fn bspatch_free_releases_zero_length_buffer() {
